LayoutTests/platform/chromium-win/fast/speech/input-text-speechbutton-expected.txt


Tests for speech button click with <input type="text" speech>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("speechInput1").value is "Pictures of the moon"
   
